Applied OOP Principle to separate GameController and UserController to adhere to SRP (please do ./gradlew clean and ./gradlew build in terminal after pulling).
